requirements.txt
Django
djangorestframework
findspark
pyspark
datetime
redis